requirements.txt
flask
python-aiml
nltk
pytholog
uvicorn
neo4j